requirements.txt
streamlit>=1.36.0
altair==4.2.2
pillow
pandas
python-pptx
requests

